com.unity.polyspatial.xr@1.2.3 ## [1.2.3] - 2024-04-23
## [1.2.2] - 2024-04-22

## [1.2.1] - 2024-04-22

## [1.2.0] - 2024-04-19

### Added
- Added a "PolySpatial XR" Plug-in Provider to XR Plug-in Managment in Project Settings.  Allows you to view ARPlane's and hands from device in editor while using Play To Device.

+++ Runtime/Input/XRTouchSpaceInteractor.cs
@@ -94,7 +94,7 @@ namespace Unity.PolySpatial.XR.Input
             }
         }
 
-        static bool TryGetInteractable(int colliderId, out XRBaseInteractable interactable)
+        static bool TryGetInteractable(long colliderId, out XRBaseInteractable interactable)
         {
             // Must get GO but seems can get collider directly at some point once PolySpatialInstanceIds of components are stored
             var go = ObjectBridge.FindObjectFromInstanceID(colliderId) as GameObject;
